cpufreq: interactive: Don't reset sampling down factor at max freq
A previous patch sets the min_sample_time to zero if the cpufreq
min/max limits have changed which would bring down the frequency
in the next sample even if min_sample_time doesnt allow it to do so.
This could severly affect benchmarks if this is applied to max
frequencies as min_sample_time for max frequencies is set to
sampling_down_factor.
This patch resets the limits_changed flag for cases where timer fires
while running at max frequencies so that sampling_down_factor is
still applicable as before.

Change-Id: I1aa305f04a7774f05c9cb21890413fd5944894f9
Signed-off-by: Rohit Gupta <rohgup@codeaurora.org>